Fixup double precision. Passed testcase.

+++ Analyzer/IIR-Filter.cc
@@ -3,7 +3,7 @@
 /*
  * 2-order IIR Filter
  */
-IIR_Filter_2ord::IIR_Filter_2ord(const float a[3], const float b[3])
+IIR_Filter_2ord::IIR_Filter_2ord(const double a[3], const double b[3])
 {
 	// setup coefficients
 	for(int i=0;i<=2;++i) {
@@ -16,7 +16,7 @@ IIR_Filter_2ord::IIR_Filter_2ord(const float a[3], const float b[3])
 
 void IIR_Filter_2ord::apply(const std::vector<double> &samples, std::vector<double> &output, int size)
 {
-	float x_1=0.0f,x_2=0.0f,y_1=0.0f,y_2=0.0f;
+	double x_1=0.0f,x_2=0.0f,y_1=0.0f,y_2=0.0f;
 
 	for (int i = 0; i < size; i++) {
 		output[i] = m_b[0]*samples[i] + m_b[1]*x_1 + m_b[2]*x_2 - m_a[1]*y_1 - m_a[2]*y_2;
